python/btcpayment/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
bit==0.4.2
celery==3.1.25
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coincurve==7.1.0
Django==2.0.2
django-celery==3.2.2
djangorestframework==3.7.7
idna==2.6
kombu==3.0.37
pycparser==2.18
pytz==2018.3
requests==2.18.4
urllib3==1.22
vine==1.1.4
